+++ changelog.md
@@ -2,6 +2,7 @@
 - Fixed wild beehives not being able to spawn in snowy areas (#56)
 - Fix fruit squeezer recipes to use tags + correct mulch rates (#59)
 - Fix Miner's bag not accepting raw ores (#58)
+- Make all biomes in `#minecraft:is_nether` tag marked as having HELLISH climate (#65)
 
 ## Forestry 1.0.5
 - Fixed wild beehives not having particle effects